Switched to BerkeleyDB for history and metadata management.

+++ library/net.c
@@ -578,7 +578,7 @@ static void dnet_schedule_command(struct dnet_net_state *st)
 		struct dnet_cmd *c = &st->rcv_cmd;
 		unsigned long long tid = c->trans & ~DNET_TRANS_REPLY;
 		dnet_log(st->n, DNET_LOG_NOTICE, "freed: size: %llu, trans: %llu, reply: %d, ptr: %p.\n",
-						c->size, tid, tid != c->trans, st->rcv_data);
+						(unsigned long long)c->size, tid, tid != c->trans, st->rcv_data);
 #endif
 		free(st->rcv_data);
 		st->rcv_data = NULL;
@@ -662,7 +662,7 @@ again:
 			}
 #if 1
 			dnet_log(n, DNET_LOG_NOTICE, "allocated: %llu, trans: %llu, reply: %d, ptr: %p.\n",
-					c->size, tid, tid != c->trans, st->rcv_data);
+					(unsigned long long)c->size, tid, tid != c->trans, st->rcv_data);
 #endif
 		}
